docs(adr): record latency decomposition delivery
